This pull request updates to OpenTelemetry version 1.5.0.

* Update dependency to OpenTelemetry 1.5
* Fix breaking changes in metrics API
* Eliminate unnecessary double to long conversions
* Simplify metric extraction for monitor

+++ opendc-simulator/opendc-simulator-compute/src/main/kotlin/org/opendc/simulator/compute/kernel/SimFairShareHypervisor.kt
@@ -77,10 +77,10 @@ public class SimFairShareHypervisor(
             if (timestamp > lastReport) {
                 listener.onSliceFinish(
                     this@SimFairShareHypervisor,
-                    (counters.demand - lastDemand).toLong(),
-                    (counters.actual - lastActual).toLong(),
-                    (counters.overcommit - lastOvercommit).toLong(),
-                    (counters.interference - lastInterference).toLong(),
+                    counters.demand - lastDemand,
+                    counters.actual - lastActual,
+                    counters.overcommit - lastOvercommit,
+                    counters.interference - lastInterference,
                     lastCpuUsage,
                     lastCpuDemand
                 )
